using customized game scene panel to enable double buffer
