+++ TowerDefense/TowerDefense/TowerDefense.csproj
@@ -57,6 +57,9 @@
     <Compile Include="GameForm.Designer.cs">
       <DependentUpon>GameForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="GameScenePanel.cs">
+      <SubType>Component</SubType>
+    </Compile>
     <Compile Include="GridParams.cs" />
     <Compile Include="Level.cs" />
     <Compile Include="Program.cs" />
